Project_9/requirements.txt
asgiref==3.8.1
black==24.8.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.8.30
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
flake8==7.1.1
idna==3.8
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
mccabe==0.7.0
more-itertools==10.5.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
pytz==2024.1
pywin32-ctypes==0.2.3
rapidfuzz==3.9.7
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.1
tomlkit==0.13.2
trove-classifiers==2024.7.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
